Add JStickyWindowSettings class for managing window settings
This commit adds the `JStickyWindowSettings` class to the `io.github.uniclog.jsticky.model` package. The `JStickyWindowSettings` class includes properties for width and height, as well as a `isModify` method for checking if the width and height values have been modified. This class is used to manage the window settings in the `JStickyData` class.